backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-Migrate==4.0.5
Flask-Caching==2.1.0

# Database
SQLAlchemy>=2.0.28
PyMySQL==1.1.0
cryptography==41.0.7
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
passlib==1.7.4
ldap3==2.9.1

# API & Utilities
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3

# AI Integration
openai==1.6.1

# File Processing
Pillow>=10.2.0
PyPDF2==3.0.1

# Task Queue (Optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0
eventlet==0.33.3